requirements.txt
pyusb
click
requests
colorama
https://github.com/pklaus/brother_ql/archive/master.zip

